# meetgraph/data/meetily/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meetily Gateway API",
    "description": "Local loopback HTTP API to drive a running Meetily desktop app. Generated; do not edit by hand.",
    "version": "1.11.0"
  },
  "paths": {
...
    "/v1/meetings": {
      "get": {
        "operationId": "get_v1_meetings",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": [
              "read"
            ]
          }
        ],
        "x-meetily-destructive": false,
        "x-meetily-required-scope": "read",
        "x-meetily-idempotent": true,
        "x-meetily-read-only": true
      }
    },
    "/v1/meetings/{id}": {
      "get": {
        "operationId": "get_v1_meetings_id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": [
              "read"
            ]
          }
        ],
        "x-meetily-destructive": false,
        "x-meetily-read-only": true,
        "x-meetily-required-scope": "read",
        "x-meetily-idempotent": true
      },
      "delete": {
        "operationId": "delete_v1_meetings_id",
        "parameters": [
          {
            "name": "delete_files",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": [
              "delete"
            ]
          }
        ],
        "x-meetily-required-scope": "delete",
        "x-meetily-read-only": false,
        "x-meetily-destructive": true,
        "x-meetily-idempotent": true
      },
      "patch": {
        "operationId": "patch_v1_meetings_id",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameMeetingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": [
              "write"
            ]
          }
        ],
        "x-meetily-read-only": false,
        "x-meetily-idempotent": false,
        "x-meetily-destructive": false,
        "x-meetily-required-scope": "write"
      }
    },
...
    "/v1/meetings/{id}/transcript": {
      "get": {
        "operationId": "get_v1_meetings_id_transcript",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": [
              "read"
            ]
          }
        ],
        "x-meetily-read-only": true,
        "x-meetily-required-scope": "read",
        "x-meetily-idempotent": true,
        "x-meetily-destructive": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RenameMeetingRequest": {
        "type": "object",
        "description": "PATCH /v1/meetings/{id} (T-12a) request body.",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}